# PERF: Series.mad