```
from asv_bench.benchmarks.stat_ops import *
self = SeriesOps()
self.setup("mad", "int")

%timeit self.time_op("mad", "int")
1.45 ms ± 22.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- master
474 µs ± 2.77 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- PR
```